Change stars to float and apply some validations

+++ data/resources.go
@@ -3,7 +3,7 @@ package data
 type Hotel struct {
 	Name        string
 	Address     string
-	Stars       int
+	Stars       float64
 	Contact     string
 	Phone       string
 	URL         string
